Add appplications submodule for cavity freezing
